test(test_api): add test_weather_parquet_integrity
After integrating the weather route and running the pipeline for 2022 and 2023, the CI tests did not work due to some problems in the parquet file (https://github.com/gregomelo/brazil_weather_data/actions/runs/7866970379/job/21462082481). This test will check if the magic bytes are at the end of the parquet file.